==> code_appendix/00_DataCollection_Cleaning.ipynb <==
# Alcohol & Substance Abuse Related Mortality Predictions

Datasets for cleaning will include:
1) IMHE County Level Mortality Rates, 1980 - 2014 (four mortality causes: alcohol use disorders, substance use disorders, self-harm, and interpersonal violence
2) IMHE County Level Alcohol Use and Prevalence Data, 2002 - 2012 (includes prevalence of alcohol use by: any, heavy & binge, as well as proportion of drinkers who are heavy drinkers and binge drinkers)
3) County Level Unemployment Data: 

Utilizing IMHE data, build a model to predict alcohol and substance abuse related mortality at the county level. 

Combine with demographics (Census/ACS) to create predictive models / explore explanatory variables. 

[CDC Wonder API Tutorial](https://www.youtube.com/watch?v=ttchrtCZ46Y)

# imports
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

import os

alcohol_death_files = os.listdir('../data/alcohol_substance_deaths/')
alcohol_prev_files = os.listdir('../data/alcohol_use_prev/')

df = pd.read_csv('../data/alcohol_substance_deaths/' + alcohol_death_files[0])
df.head(3)

## 1. Read in the IMHE Mortality Data, create dataframe
The IMHE mortality dataset came from ***LINK HERE***, and shows the mortality rate by county, year and sex (Male, Female, Both) for self-harm, alcohol use disorders, drug use disorders and interpersonal violence. 

I want to first create a function to process the whole thing. Later when I decide which to focus on, I'll create a smaller data frame that is just one of the mortality categories. 

# function to process IMHE alcohol data
def process_mortality_data(file):
    state = file.split('_')[-2].title()
    full_path = '../data/alcohol_substance_deaths/' + file


    
    #read in csv
    df = pd.read_csv(full_path, dtype={'FIPS': object})
    # drop rows that are sums of whole state
    df = df[(df['location_name'] != state) & (df['FIPS'].isna() == False)]
    
    #recase cols to save memory and so we can create a UID later
    # UID is formatted: 'FIPS-cause_id-sex_id-year_id'
    #df[['cause_id', 'sex', 'cause_name', 'location_name', 'FIPS']] = df[['location_name', 'FIPS', 'cause_id', 'sex', 'cause_name']].astype('category')
    #df['UID'] = df['FIPS'] + '-' + df['cause_id'] + '-' + df['sex_id'] + '-' + df['year_id']
    df['state'] = state
    
    # fix the state names
    # North Carolina
    df.replace('Northcarolina', 'North Carolina', inplace=True)
    df.replace('Northdakota', 'North Dakota', inplace=True)
    df.replace('Southcarolina', 'South Carolina', inplace=True)
    df.replace('Southdakota', 'South Dakota', inplace=True)
    df.replace('Newyork', 'New York', inplace=True)
    df.replace('Newjersey', 'New Jersey', inplace=True)
    df.replace('Newhampshire', 'New Hampshire', inplace=True)
    df.replace('Newmexico', 'New Mexico', inplace=True)
    df.replace('Westvirginia', 'West Virginia', inplace=True)
    df.replace('Rhodeisland', 'Rhode Island', inplace=True)
    df.replace('Dc', 'D.C.', inplace=True)
    
    df.drop(columns=['measure_name', 'cause_id', 'metric', 'sex_id', 'location_id', 'age_id', 'age_name', 'lower', 'upper'], inplace=True)
    
    #create UID columns for later joining with other dataframes
    df['UID'] = df['FIPS'] + '-' + df['year_id'].astype('string') 
    df['UID_2'] = df['FIPS'] + '-' + df['sex'] + '-' + df['year_id'].astype('string')
    
    
    return df

# process mortality data
full_mortality_df = [process_mortality_data(file) for file in alcohol_death_files]
full_mortality_df = pd.concat(full_mortality_df)
full_mortality_df.drop(columns=['measure_id', 'measure_ID'], inplace=True)

full_mortality_df['FIPS'] = full_mortality_df['FIPS'].str.zfill(5)

full_mortality_df.memory_usage(deep=True) / 1_000_000

# creating a Unique Identifier to join later with other climate dfs
subInj_df_compression_opts = dict(method='zip',
                        archive_name='alcohol_sub_inj_mortality_1980_2014.csv')

full_mortality_df.to_csv('../data/cleaned/alcohol_sub_inj_mortality_1980_2014.zip', index=False, compression=subInj_df_compression_opts)

# export full mortality df to CSV
# full_mortality_df.to_csv('data/cleaned/alcohol_sub_inj_mortality_1980_2014.csv', index=False) --> This was 146 MB, so I compressed for upload to github

#grab just self-harm
selfharm_df = full_mortality_df[full_mortality_df['cause_name'] == 'Self-harm']

#export just alcohol to csv as well
selfharm_df.to_csv('../data/cleaned/selfharm_mortality_1980_2014.csv', index=False)

#### We will also create a quick dictionary of FIPS codes that will allow us to append this data to other dataframes that might have county name and state but not the FIPS code. We need FIPS for maps & streamlit later

selfharm_df['county_state'] = selfharm_df['location_name'] + ", " + selfharm_df['state']
selfharm_unique = selfharm_df.drop_duplicates(subset=['county_state'])

# create dictionary of county name / FIPS values
#https://stackoverflow.com/questions/18012505/python-pandas-dataframe-columns-convert-to-dict-key-and-value
county_fips = dict(zip(selfharm_unique['county_state'], selfharm_unique['FIPS']))

## 2. Read in Alcohol Prevalence Data

Study Link: http://www.healthdata.org/research-article/drinking-patterns-us-counties-2002-2012 

An excel file from IMHE, tab names for reading in include:
- **Any** : "Any" drinking is defined as at least one drink of any alcoholic bevarage in the past 30 days
- **Heavy** : "Heavy" drinking is defined as the consumption, on average, of more than one drink per day for women or two drinks per day for men in the past 30 days
- **Binge** : "Binge" drinking is defined as the consumption of more than four drinks for women or five drinks for men on a single occasion at least once in the past 30 days
- **Prop. Heavy** : Proportion of All Drinkers Who Are Heavy Drinkers by year and sex, 2005-2012
- **Prop. Binge** : Proportion of All Drinkers Who Are Binge Drinkers by year and sex

Create one file that has years, counties and sex, and each column represents `any`, `heavy`, `prop_heavy`, `binge`, and `prop_binge`

def clean_alcohol_use(file, tab_name):
    df = pd.read_excel(file, tab_name)
    
    #get rid of period/space in tab_name
    #tab_name = tab_name.replace('.', '')
    #print(tab_name)
    
    # all these files have state total and national total -- we just want county totals
    # drop rows where "County" string is not in Location column
    df = df[df['Location'].str.contains('County')]
    df.reset_index(inplace=True)
    
    # drop unnecessary columns heavy and prop heavy are only 2005 to 2012, so we're dropping different columns
    if tab_name == 'Heavy':
        df.drop(columns=['index', 'Percent Change 2005-2012, Both Sexes', 'Percent Change 2005-2012, Females', 'Percent Change 2005-2012, Males'], inplace=True)
    elif tab_name == 'Prop. Heavy':
        df.drop(columns=['index', 'Percent Change 2005-2012, Both Sexes', 'Percent Change 2005-2012, Females', 'Percent Change 2005-2012, Males'], inplace=True)
    elif tab_name == 'Prop. Binge':
        df.drop(columns=['index', 'Percent Change 2002-2012, Both Sexes', 'Percent Change 2002-2012, Females', 'Percent Change 2002-2012, Males', 'Percent Change 2005-2012, Both Sexes', 'Percent Change 2005-2012, Females', 'Percent Change 2005-2012, Males'], inplace=True)
    else:
        df.drop(columns=['index', 'Percent Change 2002-2012, Both Sexes', 'Percent Change 2002-2012, Females', 'Percent Change 2002-2012, Males', 'Percent Change 2005-2012, Both Sexes', 'Percent Change 2005-2012, Females', 'Percent Change 2005-2012, Males'], inplace=True)
    
    # unmelt the columns
    # binge drinking and self-harm
    tab_name = tab_name.replace('. ', '_')
    melted_val_col_name = "alcohol_" + tab_name.lower() 
    df = df.melt(id_vars=['State', 'Location'], var_name='year_metric_category', value_name=melted_val_col_name)

    # after "melting", create new columns for year & sex
    df['year'] = df['year_metric_category'].str[0:4]
    df['sex'] = df['year_metric_category'].str[5:] 
    df.replace({'Both Sexes': 'Both', 'Males': 'Male', 'Females': 'Female'}, inplace=True)

    # create new col with county_state so we can join with another df to get the FIPS : this will be used in the streamlit app
    df['county_state'] = df['Location'] + ", " + df['State']
    
    #now add FIPS to the binge drinking data
    df['FIPS'] = df['county_state'].map(county_fips)
    
    #create UID for joining these to other dfs
    df['UID'] = df['FIPS'] + '-' + df['year'].astype('string') 
    df['UID_2'] = df['FIPS'] + '-' + df['sex'] + '-' + df['year'].astype('string')
    
    #rename some columns
    df.rename(columns={'State': 'state', 'Location': 'county_name'}, inplace=True)
    
    
    #reorder columns
    df = df[['FIPS', 'county_name', 'state', 'county_state', 'year', 'sex', 'year_metric_category', 'UID', 'UID_2',melted_val_col_name]]
    
    return df
    


alcohol_use_file = '../data/alcohol_use_prev/IHME_USA_COUNTY_ALCOHOL_USE_PREVALENCE_2002_2012_NATIONAL_Y2015M04D23.XLSX'

#read in all files using cleaned function
df_any = clean_alcohol_use(alcohol_use_file, 'Any')
df_heavy = clean_alcohol_use(alcohol_use_file, 'Heavy')
df_prop_heavy = clean_alcohol_use(alcohol_use_file, 'Prop. Heavy')
df_binge = clean_alcohol_use(alcohol_use_file, 'Binge')
df_prop_binge = clean_alcohol_use(alcohol_use_file, 'Prop. Binge')

# now we need to join files. The goal is one file that has a new column for every alcohol use category
alcohol_heavy_prop_heavy = pd.merge(df_heavy, df_prop_heavy, on='UID_2', how='left', suffixes=('', '_y'))
#drop cols
alcohol_heavy_prop_heavy.drop(columns=['FIPS_y', 'county_name_y', 'state_y', 'county_state_y', 'year_y', 'sex_y', 'year_metric_category_y', 'UID_y'], inplace=True)

#do the same to join binge/prop binge
alcohol_binge_prop_binge = pd.merge(df_binge, df_prop_binge, on='UID_2', how='left', suffixes=('', '_y'))
alcohol_binge_prop_binge.drop(columns=['FIPS_y', 'county_name_y', 'state_y', 'county_state_y', 'year_y', 'sex_y', 'year_metric_category_y', 'UID_y'], inplace=True)

alcohol_heavy_prop_heavy.memory_usage(deep=True) / 1_000_000

# the merged files are pretty big, so the big ones should be zipped
alcohol_heavy_compression_opts = dict(method='zip',
                        archive_name='alcohol_use_heavy_prop_heavy_2005_2012.csv')
alcohol_binge_compression_opts = dict(method='zip',
                        archive_name='alcohol_use_binge_prop_binge_2005_2012.csv')

#write to csv
alcohol_heavy_prop_heavy.to_csv('../data/cleaned/alcohol_use_heavy_prop_heavy_2005_2012.zip', index=False, compression=alcohol_heavy_compression_opts)
alcohol_binge_prop_binge.to_csv('../data/cleaned/alcohol_use_binge_prop_binge_2005_2012.zip', index=False, compression=alcohol_binge_compression_opts)
df_any.to_csv('../data/cleaned/alcohol_use_any_2002_2012.csv', index=False)

## 3. Unemployment Data

These had some WEIRD formatting when I downloaded them. You can see below. I opened up the file in excel, and it was just someone trying to be fancy with titles/formatting. It was ultimately quicker to correct the header issues manually in Excel and save the new csvs in a separate folder. I still had some processing to do after that my manual correction though. 

unemployment = os.listdir('../data/unemployment_excel/')

test_file = '../data/unemployment_excel/' + unemployment[0]
unemployment_test_df = pd.read_excel(test_file, header=1)

unemployment_test_df.head()

unemployment_files = os.listdir('../data/unemployment_csv/')
test_file = '../data/unemployment_csv/' + unemployment_files[0]
test_df = pd.read_csv(test_file)
test_df.head(2)

The columns for state and county FIPS area all messed up, and I need those to match up to the FIPS codes in other datasets...luckily, the FIPS code is also contained in the LAUS_CODE String. The first five digits after the CN are the FIPS code, so that's perf. I'll likely replace state FIPS with just the State name, and create a new column that also has just the county name (instead of county, state) for use later in the Streamlit app. 

Also there's lots of empty rows, and we don't need data where state is 'PR', or nan. Where ST is 'ia', we replace with DC, since this is the District of Columbia.

I'm going to write a function to clean up this data.

# function to clean the unemployment data
def clean_unemployment(file):
    full_path = '../data/unemployment_csv/' + file

    
    #read in csv
    df = pd.read_csv(full_path, dtype={'Year': 'Int64', 'unemployment_rate': 'float64' } )
    
    # drop Null rows
    df.dropna(inplace=True)
    # drop fips cols, we will get FIPS from the LAUS code
    df.drop(columns=['St_FIPS', 'County_FIPS'], inplace=True)
    
    #take care of the weird formatting in labor force, employed, unemployed columns
    df['labor_force'] = df['labor_force'].str.strip()
    df['employed'] = df['employed'].str.strip()
    df['unemployed'] = df['unemployed'].str.strip()
    
    df['labor_force'] = df['labor_force'].str.replace(',', '')
    df['employed'] = df['employed'].str.replace(',', '')
    df['unemployed'] = df['unemployed'].str.replace(',', '')
    
    # drop the N.A.
    df = df[(df['labor_force'] != 'N.A.') & (df['employed'] != 'N.A.') & (df['unemployed'] != 'N.A.')]
    
    # cast labor_force, employed, unemployed as ints
    df['labor_force'] = df['labor_force'].astype(int)
    df['employed'] = df['employed'].astype(int)
    df['unemployed'] = df['unemployed'].astype(int)
    
    #create new FIPS column
    df['FIPS'] = df['LAUS_CODE'].str[2:7]
    df['state'] = df['county_state_name'].str[-2:]
    
    df['state'].replace('ia', 'DC', inplace=True)
    
    # get rid of PR
    df = df[df['state'] != 'PR']
    df.rename(columns={'Year': 'year'}, inplace=True)
    
    # first create a UID that is 'ST-FIPS-YEAR' -- ex. 'AL-01001-1999'
    df['UID'] = df['FIPS'] + '-' + df['year'].astype('string')
    
    #reorder the columns to make it nice
    df = df[['FIPS', 'UID', 'county_state_name', 'state', 'year', 'labor_force', 'employed', 'unemployed', 'unemployment_rate']]
    
    
    return df

# test our function to see if it works
clean_unemployment(unemployment_files[14]) # it works!!

full_unemployment_df = [clean_unemployment(file) for file in unemployment_files]
full_unemployment_df = pd.concat(full_unemployment_df)

full_unemployment_df.head(2)

# write it out to a csv in the cleaned data folder
full_unemployment_df.to_csv('../data/cleaned/unemployment_1990_2019.csv', index=False)

# Combining Alcohol Use Data with Self-Harm Mortality

Data Science and Data Analysis are an interative proces. While the first portion of this notebook involved cleaning disparate datasets, now that I have decided on a problem statement, I will combine the necessary datasets to make for later visualizing and modeling.  

I will be combining the self-harm mortality by county, and the alcohol use dataframes.


# self-harm and any alcohol use
harm_alcohol_any = pd.merge(left=selfharm_df, right=df_any, on='UID_2', how='left', suffixes=('', '_y'))
harm_alcohol_any.drop(columns=['FIPS_y', 'county_name', 'state_y', 'county_state', 'year', 'sex_y', 'year_metric_category', 'UID_y'], inplace=True)

# now we need to join files. The goal is one file that has a new column for every alcohol use category
#drop cols
harm_alcohol_any

# join heavy_prop_heavy with self harm
harm_alcohol_heavy_prop_heavy = pd.merge(left=selfharm_df, right=alcohol_heavy_prop_heavy, on='UID_2', how='left', suffixes=('', '_y'))
harm_alcohol_heavy_prop_heavy.drop(columns=['FIPS_y', 'county_name', 'state_y', 'county_state', 'year', 'sex_y', 'year_metric_category', 'UID_y'], inplace=True)
harm_alcohol_heavy_prop_heavy

# now join self harm and binge_prop_binge
harm_alcohol_binge_prop_binge = pd.merge(left=selfharm_df, right=alcohol_binge_prop_binge, on='UID_2', how='left', suffixes=('', '_y'))
harm_alcohol_binge_prop_binge.drop(columns=['FIPS_y', 'county_name', 'state_y', 'county_state', 'year', 'sex_y', 'year_metric_category', 'UID_y'], inplace=True)
harm_alcohol_binge_prop_binge

# write the cleaned/joined files out to csv
harm_alcohol_any.to_csv('../data/cleaned/selfharm_alcohol_joins/selfharm_join_alcohol_any.csv', index=False)
harm_alcohol_heavy_prop_heavy.to_csv('../data/cleaned/selfharm_alcohol_joins/selfharm_join_heavy_prop_heavy.csv', index=False)
harm_alcohol_binge_prop_binge.to_csv('../data/cleaned/selfharm_alcohol_joins/selfharm_join_binge_prop_binge.csv', index=False)

# Join alcohol use, self harm and unemployment
This will be a different dataframe, since we don't have unemployment by sex

#drop the male/female rows, in favor of only both, since we don't have unemployment by gender
harm_alcohol_any_both_sexes = harm_alcohol_any[harm_alcohol_any['sex'] == 'Both']
harm_alcohol_heavy_prop_heavy_both_sexes = harm_alcohol_heavy_prop_heavy[harm_alcohol_heavy_prop_heavy['sex'] == 'Both']
harm_alcohol_binge_prop_binge_both_sexes = harm_alcohol_binge_prop_binge[harm_alcohol_binge_prop_binge['sex'] == 'Both']

# join self harm/alcohol combo dfs with unemployment
harm_alcohol_any_with_unemployment = pd.merge(harm_alcohol_any_both_sexes, full_unemployment_df, on='UID', how='inner', suffixes=('', '_y'))
harm_alcohol_heavy_with_unemployment = pd.merge(harm_alcohol_heavy_prop_heavy_both_sexes, full_unemployment_df, on='UID', how='inner', suffixes=('', '_y'))
harm_alcohol_binge_with_unemployment = pd.merge(harm_alcohol_binge_prop_binge_both_sexes, full_unemployment_df, on='UID', how='inner', suffixes=('', '_y'))

harm_alcohol_any_with_unemployment

# drop unwanted cols
harm_alcohol_any_with_unemployment.drop(columns=['FIPS_y', 'county_state_y', 'county_state_name', 'year'], inplace=True)
harm_alcohol_heavy_with_unemployment.drop(columns=['FIPS_y', 'county_state_y', 'county_state_name', 'year'], inplace=True)
harm_alcohol_binge_with_unemployment.drop(columns=['FIPS_y', 'county_state_y', 'county_state_name', 'year'], inplace=True)

# write to csv
harm_alcohol_any_with_unemployment.to_csv('../data/cleaned/selfharm_alcohol_unemployment_joins/harm_alcohol_any_with_unemployment.csv', index=False)
harm_alcohol_heavy_with_unemployment.to_csv('../data/cleaned/selfharm_alcohol_unemployment_joins/harm_alcohol_heavy_with_unemployment.csv', index=False)
harm_alcohol_binge_with_unemployment.to_csv('../data/cleaned/selfharm_alcohol_unemployment_joins/harm_alcohol_binge_with_unemployment.csv', index=False)

#### Now that the Data are cleaned, we can move on to modeling!
