==> code_appendix/02_RegressionModeling.ipynb <==
# Linear Regression Modeling


Now that I have a problem statement in hand, I can begin the modeling process. I will be creating a linear model that explores the relationships between binge drinking, sex/gender, and mortality rates by self-harm.

## Problem Statement
As the proportion of the population that binge drinks increases, does the mortality rate by self harm change? 



---
### Process

In this notebook, I'll investigate relationships between self-harm mortality rate and other variables, such as alcohol use/type, state, and sex, as well as between unemployment rate and labor force.

The models in this notebook will all be simple univariate linear models that directly compare alcohol prevalence use, sex, state, unemployment or labor force with self-harm mortality. 


# imports
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

# imports - modeling
from sklearn.model_selection import train_test_split, GridSearchCV, cross_val_score
from sklearn.preprocessing import StandardScaler, OneHotEncoder
from sklearn.linear_model import LinearRegression, Ridge, Lasso
from sklearn.ensemble import RandomForestRegressor, GradientBoostingRegressor, AdaBoostRegressor
from sklearn.metrics import mean_squared_error
from sklearn.compose import make_column_transformer, make_column_selector
from sklearn.pipeline import make_pipeline
from sklearn.feature_selection import SelectKBest

#sklearn.set_config(display='diagram')

# read in data
# 'labor_force': 'Int64', 'employed': 'float64', 'unemployed': 'float64', 'unemployment_rate': 'float64'

#read in joined selfharm_alcohol use data
harm_alcohol_any = pd.read_csv('../data/cleaned/selfharm_alcohol_joins/selfharm_join_alcohol_any.csv', dtype={'FIPS': 'object'})
harm_alcohol_heavy = pd.read_csv('../data/cleaned/selfharm_alcohol_joins/selfharm_join_heavy_prop_heavy.csv', dtype={'FIPS': 'object'})
harm_alcohol_binge = pd.read_csv('../data/cleaned/selfharm_alcohol_joins/selfharm_join_binge_prop_binge.csv', dtype={'FIPS': 'object'})

#drop the NAs
harm_alcohol_any.dropna(inplace=True)
harm_alcohol_heavy.dropna(inplace=True)
harm_alcohol_binge.dropna(inplace=True)


### First - Establishing a Baseline
To get a baseline accuracy score

harm_alcohol_binge['mx'].mean()

harm_alcohol_any

# Model 1 - Simple Univariate Regression Models
Here, I will not take into account state, county, sex or year. I will simply see if there appears to be a relationship between the different types of alcohol use/prevalence and `mx` (self-harm mortality rate). 

My success metrics will be accuracy and RMSE.

I will write a function that will allow me to try several different regression models with less code. 

# instantiate different regression models and put them in a list
lr = LinearRegression()
lasso = Lasso()
ridge = Ridge()
rf = RandomForestRegressor()
ada = AdaBoostRegressor()
gboost = GradientBoostingRegressor()

estimators = [lr, lasso, ridge, rf, ada, gboost]

# function
def model_eval(df, independent_var, dependent_var, estimator_list):
    
    #drop the "both" sexes rows, so we're left with only male and female
    df = df[df['sex'] != 'Both']
    
    X = df[[independent_var]]
    y = df[dependent_var]
    
    X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=42)
    
    for estimator in estimator_list:
        estimator.fit(X_train, y_train)

        train_score = round(estimator.score(X_train, y_train), 2)
        test_score = round(estimator.score(X_test, y_test), 2)

        rmse = round(mean_squared_error(y_test, estimator.predict(X_test), squared=False), 2)
        
        print(f'Estimator: {estimator}, Train/Test Accuracy: {train_score, test_score}, RMSE: {rmse}')
    
    return None
    

# regressions for alcohol_any
model_eval(harm_alcohol_any, 'alcohol_any', 'mx', estimators)

# regressions for alcohol_heavy
model_eval(harm_alcohol_heavy, 'alcohol_heavy', 'mx', estimators)

# regressions for alcohol_prop_heavy
model_eval(harm_alcohol_heavy, 'alcohol_prop_heavy', 'mx', estimators)

# regressions for alcohol_binge
model_eval(harm_alcohol_binge, 'alcohol_binge', 'mx', estimators)

# regressions for alcohol_binge
model_eval(harm_alcohol_binge, 'alcohol_prop_binge', 'mx', estimators)

### Simple Univariate Regression Model Summary

There doesn't appear to be much of a substantial relationship between `alcohol_any` and the self harm mortality rate, so we'll scrap that variable moving forward. 

None of the model/input variable combinations appear to show overfitting, so we can focus solely on which variables / models to include moving forward. 

`alcohol_prop_heavy` and `alcohol_prop_binge` appear to produce models with the strongest accuracy scores (although `alcohol_binge` only just barely underperforms `alcohol_prop_heavy` on measures of accuracy (except on the AdaBoostRegressor). 

Our best scoring model/variable combo used `alcohol_prop_binge` with RandomForestRegressor and the GradientBoostRegressor producing nearly identical accuracy and RMSE results. 

---

## Model 1 Optimization: Fine Tuning Our Best Simple Regression Models

Since we identified that the RandomForestRegressor and GradientBoostRegressor performed *exceptionally* well in terms of a accuracy, we will now hypertune the parameters to see if we can drive the accuracy up even higher. 


### GridSearching for params
harm_alcohol_binge_both = harm_alcohol_binge[harm_alcohol_binge['sex'] != 'Both']

X = harm_alcohol_binge_both[['alcohol_prop_binge']]
y = harm_alcohol_binge_both['mx']

X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=42)

rf = RandomForestRegressor(n_jobs=-1)
rf_params = {'criterion': ['squared_error', 'poisson'], 'max_features': ['auto', 'sqrt', 'log2'], 'verbose': [True, False]}

gboost = GradientBoostingRegressor()
gb_params = {'loss': ['squared_error', 'absolute_error', 'huber'], 'n_estimators': [100, 200], 'max_depth': [3, 5, 10]}

# write a grid search function that returns the train, test, rmse and best_params
def grid_search_eval(est, params):
    #est = estimator(n_jobs = -1)
    # instantiate model
    gs = GridSearchCV(estimator=est, param_grid=params)
    
    # fit the model
    gs.fit(X_train, y_train)
    
    # print out metrics
    print("Train Score: ", gs.score(X_train, y_train))
    print('Test Score: ', gs.score(X_test, y_test))
    print("RMSE: ", mean_squared_error(y_test, gs.predict(X_test), squared=False))
    print('Best Params: ', gs.best_params_)
    
    #return the gs instance in case we want to work with it
    return gs
    
    

# grid search for random forest
grid_search_eval(rf, rf_params)

grid_search_eval(gboost, gb_params)

lr.coef_, lr.intercept_

### Model 2 Summary: 


### Random Forest Regression with lots of vars
